+++ student.rb
@@ -1,0 +1,14 @@
+require_relative('./Person')
+
+class Student < Person
+  attr_accessor :classroom
+
+  def initialize(classroom, age, name, parent_permission)
+    @classroom = classroom
+    super(age, name, parent_permission)
+  end
+
+  def play_hooky
+    "¯\(ツ)/¯"
+  end
+end
